move fs and fs fallback into its own fs package
