+++ fs/fs.go
@@ -2,24 +2,24 @@
 
 // fs.FS is only available from go 1.16 onwards
 
-package interp
+package fs
 
 import (
-	"io/fs"
+	actualFs "io/fs"
 	"os"
 )
 
 // We use a type alias to make it easier for the pre-go1.16
 // code to fullfil this local type.
-type FS = fs.FS
+type FS = actualFs.FS
 
 // realFS complies with the fs.FS interface.
 // We use this rather than os.DirFS as DirFS has no concept of
 // what the current working directory is, whereas if we're a simple
 // passthru to os.Open then working dir is automagically taken care of.
-type realFS struct{}
+type RealFS struct{}
 
-func (dir realFS) Open(name string) (fs.File, error) {
+func (dir RealFS) Open(name string) (actualFs.File, error) {
 	f, err := os.Open(name)
 	if err != nil {
 		return nil, err
@@ -28,7 +28,7 @@ func (dir realFS) Open(name string) (fs.File, error) {
 }
 
 var (
-	ReadDir  = fs.ReadDir
-	Stat     = fs.Stat
-	ReadFile = fs.ReadFile
+	ReadDir  = actualFs.ReadDir
+	Stat     = actualFs.Stat
+	ReadFile = actualFs.ReadFile
 )
